stuff. gah I don't even remember
