+++ alchemy/sale.py
@@ -1,15 +1,5 @@
-from sqlalchemy import create_engine
-engine = create_engine('sqlite:///alchemy.db', echo=False)    # echo=True to show SQL statements
-
-from sqlalchemy.ext.declarative import declarative_base
-Base = declarative_base()
-
+from __main__ import *
 from initialize_db import Product, Order, StockTake
-
-from sqlalchemy.orm import sessionmaker
-Session = sessionmaker(bind=engine)
-session = Session()
-
 
 # show preisliste
 #	query database for last StockTake where quantity > 0
