+++ alchemy/stock_take.py
@@ -1,17 +1,7 @@
-from sqlalchemy import create_engine
-engine = create_engine('sqlite:///alchemy.db', echo=False)    # echo=True to show SQL statements
-
-from sqlalchemy.ext.declarative import declarative_base
-Base = declarative_base()
-
-from initialize_db import Product, Order, StockTake, StockTakeDetail
+from __main__ import *
 from add_products import write_products
 
 from time import time
-
-from sqlalchemy.orm import sessionmaker
-Session = sessionmaker(bind=engine)
-session = Session()
 
 
 # ask to input product to the stock take
@@ -32,7 +22,7 @@ session = Session()
 # 
 # repeat until stock take is complete
 # 
-# Leergut zusätzlich noch da?
+# Leergut zusaetzlich noch da?
 # 	Anzahl der Kasten (float, because half-crates!)
 # 	Anzahl der 0.08 Flaschen
 # 	Anzahl der 0.15 Flaschen
@@ -44,30 +34,6 @@ session = Session()
 # 	request change, return to input products, but retain order information
 # 		if changes are made, overwrite previous choice, else keep choices
 
-# this same function is defined in add_products
-def yes_no(question, yes="", no=""):
-	while True:
-		print("")
-		print(question + " j/n:")
-		choice = raw_input(">")
-		if choice == "j":
-			if yes != "":
-				print(yes)
-			return True
-		elif choice == "n":
-			if no != "":
-				print(no)
-			return False
-		else:
-			print("Bitte entweder 'j' oder 'n' eingeben")
-			continue
-
-def check_exists(input):
-	for instance in session.query(Product.artNum):
-		if instance.artNum == input:
-			return True
-	return False
-
 def edit_write_buffer(inputArtNum):
 	writeBuffer = []
 	currentProduct = session.query(Product).filter(Product.artNum == inputArtNum).one()
@@ -78,7 +44,7 @@ def edit_write_buffer(inputArtNum):
 	
 	while True:
 		try:
-			quantity = int(raw_input("Anzahl der Liefereinheiten:	")))
+			quantity = int(raw_input("Anzahl der Liefereinheiten:	"))
 		except TypeError:
 			print("Bitte nur ganze Zahlen eingeben!")
 	
@@ -89,7 +55,8 @@ def edit_write_buffer(inputArtNum):
 	
 	while True:
 		try:
-			quantity += int(raw_input("Zusaetzliche volle Flaschen:	")))
+			quantity += int(raw_input("Zusaetzliche volle Flaschen:	"))
+			break
 		except TypeError:
 			print("Bitte nur ganze Zahlen eingeben!")
 		
@@ -101,12 +68,14 @@ def edit_write_buffer(inputArtNum):
 	while True:
 		try:
 			writeBuffer.append(float(raw_input("Preis pro Liefereinheit:	")))
+			break
 		except TypeError:
 			print("Bitte nur Dezimalzahlen eingeben!")
 	
 	while True:
 		try:
 			writeBuffer.append(float(raw_input("Aufschlag pro Flasche:		")))
+			break
 		except TypeError:
 			print("Bitte nur Dezimalzahlen eingeben!")
 		
@@ -116,8 +85,16 @@ def new_stocktake_detail():
 	while True:
 		try:
 			inputArtNum = int(raw_input("Artikelnummer:		"))
+			break
 		except TypeError:
 			print("Bitte nur Ziffern eingeben!")
+	
+	if check_exists(inputArtNum) is True:
+		new_stocktake(inputArtNum)
+	else:
+		print("Artikel existiert noch nicht in der Datenbank!")
+		# TODO: pass article number to write_products
+		write_products()    # from import add_products
 	
 	writeBuffer, pfandcrates, pfandbottles = edit_write_buffer(inputArtNum)
 	if yes_no(
@@ -145,6 +122,8 @@ def extra_pfand():
 		# Anzahl der Kasten (float, because half-crates!)
 		# Anzahl der 0.08 Flaschen
 		# Anzahl der 0.15 Flaschen
+		pass
+	
 	print("Wie viele Kasten?")
 	while True:
 		try:
@@ -152,7 +131,7 @@ def extra_pfand():
 		except TypeError:
 			print("Bitte nur Ziffern eingeben!")
 	
-	while True:
+#	while True:
 		# TODO: fudge nugget, I've coded myself into a pickle.
 		# see issue #12 on github
 		
@@ -165,7 +144,7 @@ def extra_pfand():
 		)
 	)
 
-def new_stocktake(inputArtNum):
+def new_stocktake():
 	stocktake = StockTake()
 	stocktake.timestamp = time()
 	print("Notiz zu dieser Bestandaufnahme:")
@@ -173,19 +152,12 @@ def new_stocktake(inputArtNum):
 	
 	session.add(stocktake)
 	
-	while yes_no("Möchten Sie den warenbestand eines Produktes eingeben?") is True:
+	while yes_no("Moechten Sie den Warenbestand eines Produktes eingeben?") is True:
 		new_stocktake_detail()
 	
 	extra_pfand()
-	
-	session.commit()
 
 def take_stock():
 	while yes_no("Neue Bestandaufnahme eingeben?") is True:
-		inputArtNum = int(raw_input("Artikelnummer:		"))
-		if check_exists(inputArtNum) is True:
-			new_stocktake(inputArtNum)
-		elif yes_no("Artikel existiert noch nicht in der Datenbank! Moechten Sie ein neues Produkt anlegen?") is True:
-			write_products()    # from import add_products
-		else:
-			continue
+		new_stocktake()
+		session.commit()
